minor changes to geomtryUtils and move mapActivity/Callbacks
