+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2020-05-11 16:25","gitSha":"e2c22663e"}
+{"buildTime":"2020-05-12 09:59","gitSha":"8758e498e"}
