+++ app/src/main/java/org/dhis2/uicomponents/map/geometry/mapper/Mapper.kt
@@ -1,0 +1,5 @@
+package org.dhis2.uicomponents.map.geometry.mapper
+
+interface Mapper<in T, out E>{
+    fun mapTo(item: T): E
+}
