--- app/src/main/java/org/dhis2/uicomponents/others.kt
@@ -1,2 +1,0 @@
-package org.dhis2.uicomponents
-
